make uint8 on prediction

+++ deepforest/visualize.py
@@ -103,7 +103,7 @@ def plot_predictions(image, df, color=None):
     Returns:
         image: a numpy array with drawn annotations
     """    
-    image = image.copy()
+    image = image.astype("uint8").copy()
     if not color:
         if not ptypes.is_numeric_dtype(df.label):
             raise ValueError("Label column is not numeric, please convert to numeric to correctly color image {}".format(df.label.head()))
